requirements.txt
asgiref==3.8.1
attrs==23.2.0
aws-cdk-lib==2.149.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
constructs==10.3.0
Django==5.0.7
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsii==1.101.0
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
publication==0.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
